ecommerce-backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
django-countries==7.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
